Boaz/request

+++ src/application/classes/Collection.java
@@ -3,49 +3,44 @@ package application.classes;
 public class Collection {
 
 	private String name;
-	private float openseaSol;
-	private float magicEdenSol;
+	private long openseaSol;
+	private long magicEdenSol;
 	private float diff;
 	
-	public Collection(String name, float openseaSol, float magicEdenSol, float diff) {
+	public Collection(String name, long openseaSol, long magicEdenSol, float diff) {
 		super();
 		this.name = name;
 		this.openseaSol = openseaSol;
 		this.magicEdenSol = magicEdenSol;
 		this.diff = diff;
 	}
-
+	
 	public String getName() {
 		return name;
 	}
-
 	public void setName(String name) {
 		this.name = name;
 	}
-
-	public float getOpenseaSol() {
+	public long getOpenseaSol() {
 		return openseaSol;
 	}
-
-	public void setOpenseaSol(float openseaSol) {
+	public void setOpenseaSol(long openseaSol) {
 		this.openseaSol = openseaSol;
 	}
-
-	public float getMagicEdenSol() {
+	public long getMagicEdenSol() {
 		return magicEdenSol;
 	}
-
-	public void setMagicEdenSol(float magicEdenSol) {
+	public void setMagicEdenSol(long magicEdenSol) {
 		this.magicEdenSol = magicEdenSol;
 	}
-
 	public float getDiff() {
 		return diff;
 	}
-
 	public void setDiff(float diff) {
 		this.diff = diff;
 	}
+	
+
